[refactor] #1 refactor data, domain module
