+++ data/memozi/src/main/AndroidManifest.xml
