+++ data/memozi/src/main/java/com/memozi/memozi/model/remote/reponse/ResponseGetMemoziDto.kt
@@ -1,0 +1,15 @@
+package com.memozi.memozi.model.remote.reponse
+
+import com.memozi.memozi.domain.model.Memozi
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ResponseGetMemoziDto(
+    @SerialName("test")
+    val test: String
+)
+
+fun ResponseGetMemoziDto.toDomain() = Memozi(
+    test = this.test
+)
