+++ data/memozi/src/main/java/com/memozi/memozi/model/remote/request/RequestPostMemoziDto.kt
@@ -1,0 +1,15 @@
+package com.memozi.memozi.model.remote.request
+
+import com.memozi.memozi.domain.model.Memozi
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class RequestPostMemoziDto(
+    @SerialName("test")
+    val test: String
+)
+
+fun Memozi.toData() = RequestPostMemoziDto(
+    test = this.test
+)
