+++ data/memozi/src/main/java/com/memozi/memozi/source/local/MemoziLocalDataSource.kt
@@ -1,0 +1,9 @@
+package com.memozi.memozi.source.local
+
+import com.memozi.datastore.MemoziLocalData
+import kotlinx.coroutines.flow.Flow
+
+interface MemoziLocalDataSource {
+    val memoziLocalData: Flow<MemoziLocalData>
+    suspend fun setMemoziLocalData(memoziData: MemoziLocalData)
+}
