+++ data/memozi/src/main/java/com/memozi/memozi/source/remote/MemoziRemoteDataSource.kt
@@ -1,0 +1,10 @@
+package com.memozi.memozi.source.remote
+
+import com.memozi.memozi.model.remote.reponse.ResponseGetMemoziDto
+import com.memozi.memozi.model.remote.request.RequestPostMemoziDto
+import com.memozi.network.model.BaseResponse
+
+interface MemoziRemoteDataSource {
+    suspend fun getMemozi(): BaseResponse<ResponseGetMemoziDto>
+    suspend fun postMemozi(requestPostMemoziDto: RequestPostMemoziDto): BaseResponse<Unit>
+}
