--- data/src/main/java/com/memozi/data/model/remote/request/RequestPostMemoziDto.kt
@@ -1,15 +1,0 @@
-package com.memozi.data.model.remote.request
-
-import com.memozi.domain.model.Memozi
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class RequestPostMemoziDto(
-    @SerialName("test")
-    val test: String
-)
-
-fun Memozi.toData() = RequestPostMemoziDto(
-    test = this.test
-)
