+++ domain/memozi/src/main/AndroidManifest.xml
