+++ domain/memozi/src/main/java/com/memozi/memozi/domain/model/Memozi.kt
@@ -1,0 +1,5 @@
+package com.memozi.memozi.domain.model
+
+data class Memozi(
+    val test: String
+)
