+++ domain/memozi/src/main/java/com/memozi/memozi/domain/repository/MemoziRepository.kt
@@ -1,0 +1,11 @@
+package com.memozi.memozi.domain.repository
+
+import com.memozi.memozi.domain.model.Memozi
+import com.memozi.model.CoreData
+import kotlinx.coroutines.flow.Flow
+
+interface MemoziRepository {
+    suspend fun getMemozi(): Result<Memozi>
+    suspend fun postMemozi(memozi: Memozi): Result<Unit>
+    fun getData(): Flow<CoreData>
+}
