+++ domain/memozi/src/main/java/com/memozi/memozi/domain/usecase/GetMemoziUseCase.kt
@@ -1,0 +1,10 @@
+package com.memozi.memozi.domain.usecase
+
+import com.memozi.memozi.domain.repository.MemoziRepository
+import javax.inject.Inject
+
+class GetMemoziUseCase @Inject constructor(
+    private val memoziRepository: MemoziRepository
+) {
+    suspend operator fun invoke() = memoziRepository.getMemozi()
+}
