+++ domain/memozi/src/main/java/com/memozi/memozi/domain/usecase/PostMemoziUseCase.kt
@@ -1,0 +1,11 @@
+package com.memozi.memozi.domain.usecase
+
+import com.memozi.memozi.domain.model.Memozi
+import com.memozi.memozi.domain.repository.MemoziRepository
+import javax.inject.Inject
+
+class PostMemoziUseCase @Inject constructor(
+    private val memoziRepository: MemoziRepository
+) {
+    suspend operator fun invoke(memozi: Memozi) = memoziRepository.postMemozi(memozi)
+}
