--- domain/src/main/java/com/memozi/domain/model/Memozi.kt
@@ -1,5 +1,0 @@
-package com.memozi.domain.model
-
-data class Memozi(
-    val test: String
-)
