+++ local/memozi/build.gradle.kts
@@ -8,6 +8,6 @@ android {
 dependencies {
     implementation(projects.core.model)
     implementation(projects.core.datastore)
-    implementation(projects.data)
+    implementation(projects.data.memozi)
     implementation(libs.bundles.datastore)
 }
